backend/requirements.txt
Django>=5.0
djangorestframework
django-cors-headers
psycopg2-binary
dj-database-url
python-dotenv
whitenoise
gunicorn
